Move new type to it's own file
